+++ OpenSauce/Halo1/Halo1_CE/Interface/OpenSauceUI/GwenUI/c_control_gwen.cpp
@@ -38,7 +38,12 @@ namespace Yelo
 		{
 			auto& bounds = m_control->GetBounds();
 
-			return rectangle2d { bounds.x, bounds.y, bounds.h, bounds.w };
+			return rectangle2d { 
+				static_cast<int16>(bounds.x),
+				static_cast<int16>(bounds.y),
+				static_cast<int16>(bounds.h),
+				static_cast<int16>(bounds.w)
+			};
 		}
 
 		void* c_control_gwen::GetControlPtr() const
